Moves integrations to dedicated file
