+++ SNData/__init__.py
@@ -6,8 +6,11 @@ surveys.
 """
 
 from . import csp, des, sdss
-from ._utils import parse_snoopy_data, query_ned_coords
-from  ._osc import query_osc_photometry, query_osc_spectra
+from ._integrations import parse_snoopy_data
+from ._integrations import query_ned_coords
+from ._integrations import query_osc
+from ._integrations import query_osc_photometry
+from ._integrations import query_osc_spectra
 
 __version__ = '0.0.8'
 __author__ = 'Daniel Perrefort'
