+++ SNData/_utils.py
@@ -7,72 +7,8 @@ import tarfile
 from pathlib import Path, PosixPath
 from tempfile import TemporaryFile
 
-import numpy as np
 import requests
-import sncosmo
-from astropy.table import Table
 from tqdm import tqdm
-
-
-def parse_snoopy_data(path):
-    """Return data from a snoopy file as an astropy table
-
-    Args:
-        path (str): The file path of a snoopy input file
-
-    Returns:
-        An astropy table with columns 'time', 'band', 'mag', and 'mag_err'
-    """
-
-    out_table = Table(
-        names=['time', 'band', 'mag', 'mag_err'],
-        dtype=[float, object, float, float]
-    )
-
-    with open(path) as ofile:
-        # Get meta data from first line
-        name, z, ra, dec = ofile.readline().split()
-        out_table.meta['obj_id'] = name
-        out_table.meta['redshift'] = float(z)
-        out_table.meta['ra'] = float(ra)
-        out_table.meta['dec'] = float(dec)
-
-        # Read photometric data from the rest of the file
-        band = None
-        for line in ofile.readlines():
-            line_list = line.split()
-            if line.startswith('filter'):
-                band = line_list[1]
-                continue
-
-            time, mag, mag_err = line_list
-            out_table.add_row([time, band, mag, mag_err])
-
-    return out_table
-
-
-def register_filter(file_path, filt_name, force=False):
-    """Registers filter profiles with sncosmo if not already registered
-
-    Args:
-        file_path (str): Path of an ascii table with wavelength (Angstrom)
-                          and transmission columns
-        filt_name (str): The name of the registered filter.
-        force    (bool): Whether to re-register a band if already registered
-    """
-
-    # Get set of registered builtin and custom bandpasses
-    available_bands = set(
-        k[0] for k in sncosmo.bandpasses._BANDPASSES._loaders)
-    available_bands.update(
-        k[0] for k in sncosmo.bandpasses._BANDPASSES._instances)
-
-    # Register the new bandpass
-    if filt_name not in available_bands:
-        filt_data = np.genfromtxt(file_path).T
-        band = sncosmo.Bandpass(filt_data[0], filt_data[1])
-        band.name = filt_name
-        sncosmo.register(band, force=force)
 
 
 def download_file(url, out_file):
@@ -141,36 +77,3 @@ def build_pbar(data, verbose):
         iter_data = data
 
     return iter_data
-
-
-def query_ned_coords(uia_name):
-    """Return the J2000 RA and Dec for UIA named supernovae from NED
-
-    Args:
-        uia_name (str): SN name (e.g. ['SN2011fe'])
-
-    Returns:
-        RA position in degrees
-        Dec position in degrees
-    """
-
-    if not uia_name.lower().startswith('sn'):
-        uia_name = "SN" + uia_name
-
-    url = (
-        f"http://ned.ipac.caltech.edu/cgi-bin/objsearch?objname={uia_name}"
-        "&out_csys=Equatorial"
-        "&out_equinox=J2000.0"
-        "&of=ascii_bar"
-        "&list_limit=5"
-        "&img_stamp=NO"
-    )
-
-    response = requests.get(url)
-    response.raise_for_status()
-
-    obj_data = response.content.decode('utf-8').split('\n')[-2].split('|')
-    if obj_data[0] != '1':
-        raise RuntimeError(f"Could not retreive coordinates for {uia_name}")
-
-    return float(obj_data[2]), float(obj_data[3])
