+++ SNData/des/sn3yr/_data_parsing.py
@@ -8,6 +8,7 @@ from astropy.table import Table
 
 from . import _meta as meta
 from ._data_download import _raise_for_data
+from ... import _integrations as integrations
 from ... import _utils as utils
 
 
@@ -21,16 +22,17 @@ def register_filters(force=False):
     _raise_for_data()
     for _file_name, _band_name in zip(meta.filter_file_names, meta.band_names):
         fpath = meta.filter_dir / _file_name
-        utils.register_filter(fpath, _band_name, force=force)
+        integrations.register_filter(fpath, _band_name, force=force)
 
 
 def get_available_tables():
     """Get numbers of available tables for this survey / data release"""
 
+    # _raise_for_data()
     raise RuntimeError('No Vizier tables available for this paper.')
-    _raise_for_data()
 
 
+# noinspection PyUnusedLocal
 def load_table(table_num):
     """Load a table from the data paper for this survey / data
 
@@ -38,8 +40,8 @@ def load_table(table_num):
         table_num (int): The published table number
     """
 
+    # _raise_for_data()
     raise RuntimeError('No Vizier tables available for this paper.')
-    _raise_for_data()
 
 
 def get_available_ids():
@@ -127,7 +129,7 @@ def iter_data(verbose=False, format_sncosmo=False, filter_func=None):
 
     Args:
         verbose  (bool, dict): Optionally display progress bar while iterating
-        format_sncosmo (bool): Format data for use with SNCosmo (Default: False)
+        format_sncosmo (bool): Format data for SNCosmo.fit_lc (Default: False)
         filter_func    (func): An optional function to filter outputs by
 
     Yields:
@@ -147,4 +149,3 @@ def iter_data(verbose=False, format_sncosmo=False, filter_func=None):
 
         if filter_func(data_table):
             yield data_table
-
